// services/madoc-ts/src/schemas/collection-full.ts
import { InternationalString, MetadataItem } from '@hyperion-framework/types';
import { Pagination } from './_pagination';

export type CollectionFull = {
  collection: {
    id: number;
    thumbnail: null | string;
    label: InternationalString;
    metadata?: Array<MetadataItem>;
    requiredStatement?: MetadataItem;
    summary?: InternationalString;
    items: Array<{
      id: number;
      label: InternationalString;
      thumbnail: null | string;
      canvasCount?: number;
    }>;
  };
  pagination: Pagination;
};